% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_legend.R
\name{extract_legend}
\alias{extract_legend}
\title{extract the legend of a \code{ggplot} object}
\usage{
extract_legend(plot)
}
\arguments{
\item{plot}{a \code{ggplot} object}
}
\description{
extract the legend of a \code{ggplot} object
}
