% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_impute.R
\name{impute_variable}
\alias{impute_variable}
\title{imputes only one variable with knn}
\usage{
impute_variable(var, my_data, my_data_impute, k)
}
\arguments{
\item{var}{a string indicating the variable to be imputed}

\item{my_data}{a tibble / data.frame containing the dataset with missing values}

\item{my_data_impute}{a tibble / data.frame containing the dataset with filled missing values}

\item{k}{the number of neighbors}
}
\description{
imputes only one variable with knn
}
