% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_scales.R
\name{inverse_minus_log}
\alias{inverse_minus_log}
\title{Converts positive values to values between 0 and 1}
\usage{
inverse_minus_log(x, base = exp(1))
}
\arguments{
\item{x}{a numeric value}

\item{base}{base used for log}
}
\value{
a numeric value
}
\description{
Converts positive values to values between 0 and 1
}
