% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_impute.R
\name{knn_impute_init}
\alias{knn_impute_init}
\title{init of knn impute, if the variable is numeric then imputes with mean of all values
otherwise takes the most repeated levels.}
\usage{
knn_impute_init(var)
}
\arguments{
\item{var}{a variable with missing values as NA}
}
\value{
the var vector with imputed values
}
\description{
init of knn impute, if the variable is numeric then imputes with mean of all values
otherwise takes the most repeated levels.
}
