% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_scales.R
\name{minus_log}
\alias{minus_log}
\title{Convert values between 0 and 1 to log scales and changes its sign}
\usage{
minus_log(x, base = exp(1))
}
\arguments{
\item{x}{a numeric value}

\item{base}{base used for log}
}
\value{
a numeric value
}
\description{
Convert values between 0 and 1 to log scales and changes its sign
}
