% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_scales.R
\name{minus_log_trans}
\alias{minus_log_trans}
\title{scales to extend small numeric values}
\usage{
minus_log_trans(base = exp(1))
}
\arguments{
\item{base}{the base of the logarithm}
}
\description{
scales to extend small numeric values
}
