% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_impute.R
\name{most_repeated}
\alias{most_repeated}
\title{returns the most repeated value in var}
\usage{
most_repeated(var)
}
\arguments{
\item{var}{a vector}
}
\value{
the most repeated value
}
\description{
returns the most repeated value in var
}
