% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\description{
ggplot objects can be passed in ..., or to plotlist (as a list of ggplot objects)
 @param cols number of columns in layout
 @param layout a matrix specifying the layout. If present, \code{cols} is
 ignored.
 If the layout is something like \code{matrix(c(1,2,3,3), nrow=2, byrow=TRUE)},
 # then plot 1 will go in the upper left, 2 will go in the upper right, and
 # 3 will go all the way across the bottom.
 @export
 taken from https://support.bioconductor.org/p/68650/
 link to highlight axis too http://rpubs.com/lgadar/matrix-visualizations
}
