% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_venn.R
\name{plot_venn}
\alias{plot_venn}
\title{plot_venn function}
\usage{
plot_venn(venn, colors)
}
\arguments{
\item{venn}{a venn diagram generated with compute.Venn function}

\item{colors}{a vector of the same length as sets used to generated the venn object}
}
\value{
a plot with the venn diagram
}
\description{
A wrapper of the plot Vennerable function to change colors, and white fill
}
\examples{
\dontrun{

v1 = lit_snps \%>\% map(as.character) \%>\%
    Venn() \%>\% compute.Venn(doWeights = FALSE)

plot_venn(v1,c("red","black"))

}
}
