% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_functions.R
\name{remove_names}
\alias{remove_names}
\title{Removes the names of the vector}
\usage{
remove_names(x)
}
\arguments{
\item{x}{Vector to remove names}
}
\description{
\code{remove_names} is a function to remove the names of a vector.
}
\examples{
x = c("a" = 1,"b" = 3)
remove_names(x)

}
