% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_list.R
\name{save_list}
\alias{save_list}
\title{saves a list of plots with the same parameters}
\usage{
save_list(filenames, plots, ...)
}
\arguments{
\item{filenames}{a vector with the filenames for the plots}

\item{plots}{a list of plots of the same length than \code{filenames}}
}
\value{
nothing but it saves a bunch of plots
}
\description{
saves a list of plots with the same parameters
}
