% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_as_matrix.R
\name{scale_as_matrix}
\alias{scale_as_matrix}
\title{scale_as_matrix}
\usage{
scale_as_matrix(data, key_col, value_col, scale = TRUE, center = TRUE)
}
\arguments{
\item{data}{a tibble containing a matrix of data}

\item{key_col}{Strings giving names of key and value columns to create}

\item{scale}{either a logical value or a numeric-alike vector of length equal to the number of columns of \code{data} - 1.}

\item{center}{either a logical value or numeric-alike vector of length equal to the number of columns in \code{data} - 1, where `numeric-alike` means that \code{as.numeric(.)} will be appled successfully is \code{is.numeric(.)} is not true.}
}
\value{
A tibble with the same shape as \code{data}, with the columns scaled.
}
\description{
Adapted version of \code{scale} for tibble objects
}
