% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_functions.R
\name{stri_subset}
\alias{stri_subset}
\title{Obtains the subset of strings that don't have the pattern}
\usage{
stri_subset(string, pattern)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something coercible to one.}

\item{pattern}{Pattern to look for.}
}
\description{
\code{stri_subset} is a wrapper of \code{str_subset} that obtains the
subset of strings that don't contain the pattern
}
\examples{

 fruit <- c("apple", "banana", "pear", "pinapple")
 stri_subset(fruit, "p")

}
