% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_black.R
\name{theme_black}
\alias{theme_black}
\title{theme_black}
\usage{
theme_black(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{font size}

\item{base_family}{font to use}
}
\value{
a ggplot theme
}
\description{
a dark theme for ggplot
}
\examples{
ggplot(mtcars, aes(wt, mpg)) + geom_point()
ggplot(mtcars, aes(wt, mpg)) + geom_point(color = "white") + theme_black()
}
