% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{load2env}
\alias{load2env}
\title{Tidy load from file to list}
\usage{
load2env(file, env = new.env())
}
\arguments{
\item{file}{string with the name of an RData file}

\item{env}{an \code{environment}, to which the file is going to be load into. By default the function creates a new one.}
}
\description{
To avoid use commands of the form \code{load(file)}, this function
loads the file into an environment and then return the environment.
}
\examples{

\dontrun{

load2env(file)

}
}
